requirements.txt
# Core packages
flask
flask-cors
python-dotenv
google-genai
groq
gunicorn
